refactor(tasks): split the four files the outcome model pushed over 800
The outcome work took tasks.py, tasks_conversation.py, task_runner_auto.py
and task_runner_exec.py past the per-file limit. Split by cohesion rather
than shuffling lines to fit, following the same rule as #116:

- app/task_finalize.py — `finalize_terminal_state` (340 lines) moves out
  of task_runner_auto. It belongs beside task_outcome: that module decides
  WHAT a run produced, this one turns the verdict into a transition, an
  SSE event and a commit. It was never specific to the auto pipeline —
  the follow-up pipeline calls it too. 818 -> 478.

- app/routers/task_submission.py — the submission/verdict pair
  (`retain_submission`, `refuse`, `declared_gaps`, `SetTaskResultRequest`)
  moves out of routers/tasks.py as one unit. 842 -> 794.

- `clear_run_state()` joins task_outcome, replacing three hand-written
  copies of "null every resolver input" in tasks_conversation. The module
  that owns those inputs now owns clearing them, so a future input can't
  be added to the resolver and forgotten at one of the three sites.
  819 -> 799.

- The two byte-identical terminal blocks in task_runner_exec collapse
  into `_fail_on_agent_error`. 802 -> 795.

No behaviour change. Tests repointed at the new homes: the two finalizer
unit tests patched `task_runner_auto.{async_session,agent_manager}` and
now patch `task_finalize`, and workflow conftest registers the new module
in both patch lists — without that the finalizer ran against the real
session maker and 49 workflow tests failed.

docs/tasks.md documents the outcome model the doc bot flagged as missing:
the four columns and who writes each, the precedence, why INCOMPLETE
completes, retain-before-review, the `incomplete` exit, and the
single terminal-state writer.
