+++ app/routers/task_submission.py
@@ -1,0 +1,75 @@
+"""Submission and verdict — the two halves of ``set_task_result``.
+
+A reviewer gate refuses by raising, and the raise used to happen before
+``task.result`` was assigned, so a refusal DESTROYED the submission: N
+refusals left the run holding nothing and the end-of-turn fallback had
+only chat prose to offer. What the agent submitted and what the review
+said about it are separate facts, recorded separately here, and only a
+submission that clears every gate becomes ``accepted_result``.
+
+Split out of ``routers/tasks.py`` (over the per-file line limit) as one
+cohesive unit rather than by cutting lines to fit.
+"""
+
+from datetime import UTC, datetime
+import json
+
+from fastapi import HTTPException
+from pydantic import BaseModel
+from sqlalchemy.ext.asyncio import AsyncSession
+
+from app.models.task import Task
+
+
+class SetTaskResultRequest(BaseModel):
+    result: str
+    # The agent's own account of what it could NOT finish. Does not
+    # bypass a single gate — the submission is reviewed exactly as
+    # before. What it buys is an honest way to say "here is what I got,
+    # here is what is missing" without reaching for set_task_error,
+    # which means unrecoverable failure and was being used as an
+    # escape hatch because it was the only door out.
+    incomplete: list[str] | None = None
+
+
+async def retain_submission(db: AsyncSession, task: Task, text: str) -> None:
+    """Record what the agent submitted, before any gate can refuse it.
+
+    Committed immediately so the submission survives the ``raise`` that
+    a refusal is delivered by.
+    """
+    task.submitted_result = text
+    task.submission_count = (task.submission_count or 0) + 1
+    task.updated_at = datetime.now(UTC).isoformat()
+    await db.commit()
+
+
+def declared_gaps(body: SetTaskResultRequest) -> tuple[str, ...]:
+    """Agent-declared unfinished items, normalised."""
+    return tuple(
+        str(g).strip() for g in (body.incomplete or ()) if str(g).strip()
+    )
+
+
+async def refuse(
+    db: AsyncSession,
+    task: Task,
+    reason: str,
+    gaps: tuple[str, ...] = (),
+) -> None:
+    """Annotate the retained submission with this refusal, then 400.
+
+    Never returns. The gaps land on the task so that if the run ends
+    without an accepted submission, the finalizer can still ship the
+    best attempt with an honest list of what it is missing rather than
+    nothing at all.
+    """
+    # A gate with no natural item list still has to leave a trace: if
+    # this run never gets an accepted submission, these gaps are the
+    # only account of WHY the retained deliverable is partial. Fall back
+    # to the prose, capped so one refusal can't swamp the result.
+    recorded = list(gaps) or [reason[:500]]
+    task.review_gaps = json.dumps(recorded, ensure_ascii=False)
+    task.updated_at = datetime.now(UTC).isoformat()
+    await db.commit()
+    raise HTTPException(status_code=400, detail=reason)
