+++ app/task_runner_exec.py
@@ -51,6 +51,37 @@ _send_task_completed = send_task_completed
 _send_task_failed = send_task_failed
 
 
+async def _fail_on_agent_error(db, task, task_id, outcome) -> bool:
+    """Land the task in FAILED if the outcome says so. True if it did.
+
+    Only an outcome with no deliverable at all — or an infra-detected
+    error — is terminal here; an agent-reported error over real output
+    was already folded into the result as a caveat by
+    ``apply_outcome``. Shared by both cleanup paths in this module,
+    which were byte-identical copies.
+    """
+    if not (outcome.kind is OutcomeKind.FAILED and task.error):
+        return False
+    assert_transition(task.status, TaskStatus.FAILED)
+    task.status = TaskStatus.FAILED
+    task.updated_at = datetime.now(UTC).isoformat()
+    await db.commit()
+    await event_bus.emit(
+        'task_update',
+        {
+            'task_id': task_id,
+            'status': TaskStatus.FAILED,
+            'error': task.error,
+        },
+    )
+    _send_task_failed(
+        task,
+        phase=TaskFailurePhase.RUNNING,
+        category=TaskFailureCategory.AGENT_SET_ERROR,
+    )
+    return True
+
+
 async def _fail_task_external_config_override(
     task_id: str, ext_err: ExternalConfigOverrideError
 ) -> None:
@@ -347,28 +378,9 @@ async def execute_planned_task(task_id: str, store: Store | None):
                     )
                     return
 
-                # Terminal verdict from the resolved outcome. An agent
-                # error over real output is a caveat, not a failure —
-                # only an outcome with no deliverable (or an
-                # infra-detected error) lands in FAILED.
-                if outcome.kind is OutcomeKind.FAILED and task.error:
-                    assert_transition(task.status, TaskStatus.FAILED)
-                    task.status = TaskStatus.FAILED
-                    task.updated_at = datetime.now(UTC).isoformat()
-                    await db.commit()
-                    await event_bus.emit(
-                        'task_update',
-                        {
-                            'task_id': task_id,
-                            'status': TaskStatus.FAILED,
-                            'error': task.error,
-                        },
-                    )
-                    _send_task_failed(
-                        task,
-                        phase=TaskFailurePhase.RUNNING,
-                        category=TaskFailureCategory.AGENT_SET_ERROR,
-                    )
+                # Terminal verdict from the resolved outcome (an agent
+                # error over real output is a caveat, not a failure).
+                if await _fail_on_agent_error(db, task, task_id, outcome):
                     return
 
                 # Check if agent exited with an error result (CLI
@@ -681,28 +693,9 @@ async def execute_woken_task(task_id: str, store: Store | None):
                     )
                     return
 
-                # Terminal verdict from the resolved outcome. An agent
-                # error over real output is a caveat, not a failure —
-                # only an outcome with no deliverable (or an
-                # infra-detected error) lands in FAILED.
-                if outcome.kind is OutcomeKind.FAILED and task.error:
-                    assert_transition(task.status, TaskStatus.FAILED)
-                    task.status = TaskStatus.FAILED
-                    task.updated_at = datetime.now(UTC).isoformat()
-                    await db.commit()
-                    await event_bus.emit(
-                        'task_update',
-                        {
-                            'task_id': task_id,
-                            'status': TaskStatus.FAILED,
-                            'error': task.error,
-                        },
-                    )
-                    _send_task_failed(
-                        task,
-                        phase=TaskFailurePhase.RUNNING,
-                        category=TaskFailureCategory.AGENT_SET_ERROR,
-                    )
+                # Terminal verdict from the resolved outcome (an agent
+                # error over real output is a caveat, not a failure).
+                if await _fail_on_agent_error(db, task, task_id, outcome):
                     return
 
                 # Check if agent exited with an error result (CLI
