+++ tests/unit/test_finalizer_interactive_qa.py
@@ -39,7 +39,7 @@ from sqlalchemy.pool import StaticPool
 import app.database as _db
 from app.models.base import Base
 from app.models.task import Task
-import app.task_runner_auto as _auto
+import app.task_finalize as _fin
 
 pytestmark = pytest.mark.unit
 
@@ -68,9 +68,9 @@ async def env(monkeypatch):
         engine, class_=AsyncSession, expire_on_commit=False
     )
     monkeypatch.setattr(_db, 'async_session', maker)
-    monkeypatch.setattr(_auto, 'async_session', maker)
+    monkeypatch.setattr(_fin, 'async_session', maker)
     fake_mgr = _FakeManager()
-    monkeypatch.setattr(_auto, 'agent_manager', fake_mgr)
+    monkeypatch.setattr(_fin, 'agent_manager', fake_mgr)
     yield maker, fake_mgr
 
 
@@ -122,7 +122,7 @@ class TestFinalizerInteractiveQA:
             _asked_user_question=True,
             _tool_use_since_answer=False,
         )
-        await _auto._finalize_terminal_state(task_id, my_session=sess)
+        await _fin.finalize_terminal_state(task_id, my_session=sess)
         async with maker() as db:
             t = await db.get(Task, task_id)
             assert t.status == 'waiting', (
@@ -145,7 +145,7 @@ class TestFinalizerInteractiveQA:
             _tool_use_since_answer=True,
         )
         sess = mgr.get_session(task_id)
-        await _auto._finalize_terminal_state(task_id, my_session=sess)
+        await _fin.finalize_terminal_state(task_id, my_session=sess)
         async with maker() as db:
             t = await db.get(Task, task_id)
             assert t.status == 'completed', (
@@ -168,7 +168,7 @@ class TestFinalizerInteractiveQA:
             _tool_use_since_answer=False,
         )
         sess = mgr.get_session(task_id)
-        await _auto._finalize_terminal_state(task_id, my_session=sess)
+        await _fin.finalize_terminal_state(task_id, my_session=sess)
         async with maker() as db:
             t = await db.get(Task, task_id)
             assert t.status == 'completed', (
@@ -188,7 +188,7 @@ class TestFinalizerInteractiveQA:
             _is_error_result=False, _agent_success=True
         )
         sess = mgr.get_session(task_id)
-        await _auto._finalize_terminal_state(task_id, my_session=sess)
+        await _fin.finalize_terminal_state(task_id, my_session=sess)
         async with maker() as db:
             t = await db.get(Task, task_id)
             assert t.status == 'completed'
